rename ucase module to casemap
